Refactor ConnectionSupplier
