+++ src/main/java/com/github/kagkarlsson/jdbc/ExternallyManagedConnection.java
@@ -1,0 +1,42 @@
+/**
+ * Copyright (C) Gustav Karlsson
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *         http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.github.kagkarlsson.jdbc;
+
+import java.sql.Connection;
+
+public class ExternallyManagedConnection implements ConnectionSupplier {
+
+    private final Connection externallyManagedConnection;
+
+    public ExternallyManagedConnection(Connection externallyManagedConnection) {
+        this.externallyManagedConnection = externallyManagedConnection;
+    }
+
+    @Override
+    public Connection getConnection() {
+        return externallyManagedConnection;
+    }
+
+    @Override
+    public boolean commitWhenAutocommitDisabled() {
+        return false;
+    }
+
+    @Override
+    public boolean isExternallyManagedConnection() {
+        return true;
+    }
+}
